requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
